move options for app to class of static methods
