+++ tests/test_units/test_app_utils.py
@@ -1,8 +1,4 @@
-from SOPRANO.utils.app_utils import (
-    get_annotated_input_options,
-    get_coordinate_options,
-    get_immunopeptidome_options,
-)
+from SOPRANO.utils.app_utils import AppOptions
 from SOPRANO.utils.path_utils import Directories
 
 
@@ -27,17 +23,23 @@ def _check_detected(dir_method, options_method, ext):
 
 def test_get_annotated_input_options():
     _check_detected(
-        Directories.app_annotated_inputs, get_annotated_input_options, "anno"
+        Directories.app_annotated_inputs,
+        AppOptions.get_annotated_input_options,
+        "anno",
     )
 
 
 def test_get_immunopeptidome_options():
     _check_detected(
-        Directories.app_immunopeptidomes, get_immunopeptidome_options, "bed"
+        Directories.app_immunopeptidomes,
+        AppOptions.get_immunopeptidome_options,
+        "bed",
     )
 
 
 def test_get_coordinate_options():
     _check_detected(
-        Directories.app_coordinate_files, get_coordinate_options, "bed"
+        Directories.app_coordinate_files,
+        AppOptions.get_coordinate_options,
+        "bed",
     )
